backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.1.0
rq==1.16.2
gitpython==3.1.43
httpx==0.27.2
python-multipart==0.0.12
